[ESM] Convert the `external/builder/`-folder to use standard modules
